api/requirements.txt
# requirements.txt
Flask==2.0.3
google-cloud-firestore==2.6.0
google-cloud-secret-manager==2.9.0
gunicorn==20.1.0